Split code
Split code into several files.
Format with Black.
Align with PEP guidelines.
