+++ CHANGES.md
@@ -1,0 +1,6 @@
+# Change Log
+
+## 0.5.0
+* Build DOCX templates
+* Provide a basic loader
+* Allows for dynamic display of project name
