+++ README.md
@@ -1,5 +1,4 @@
-# TODO:
-+ Create a front-end
+[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 
 # Introduction
 This project is written in Python and aim to automatically fill-in a CV template, based on a given JSON file.
