+++ cv/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+__init__.py (cv)
+Author: Gilson, K
+"""
+
+from .education import Education
+from .employee import Employee
+from .language import Language
+from .project import Project
+from .work_experience import WorkExperience
