+++ gui/__init__.py
@@ -1,0 +1,9 @@
+# -* coding: utf-8 -*-
+"""
+__init__.py (gui)
+Author: Gilson, K
+"""
+
+from .control import ControlFrame
+from .load_files import LoadFilesFrame
+from .projects_list import ProjectsListFrame
